feat(api)! Get question by id TRELLO: 10

+++ src/main/java/com/forum/app/repository/TopicRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.forum.app.entity.Topic;
 
-public interface TopicRepository extends JpaRepository<Topic, Integer> {
+public interface TopicRepository extends JpaRepository<Topic, Long> {
 }
